+++ src/correlate_t3.h
@@ -19,7 +19,7 @@ int under_max_distance_t3(t3_t *left, t3_t *right, options_t *options);
 int over_min_distance_t3(t3_t *left, t3_t *right, options_t *options);
 
 int correlate_t3(FILE *in_stream, FILE *out_stream, options_t *options);
-int correlate_t3_block(FILE *out_stream, long long *record_number,
+int correlate_t3_block(FILE *out_stream, int64_t *record_number,
 		t3_queue_t *queue,
 		permutations_t *permutations,
 		offsets_t *offsets, t3_t *correlation_block,
